Add codility dir
