+++ src/main/java/com/codility/decimal_zip/Solution.java
@@ -1,0 +1,6 @@
+package com.codility.decimal_zip;
+
+public interface Solution 
+{
+    public int solution(int A, int B);
+}
