--- src/main/java/com/example/DecimalZipProblem/Solution.java
@@ -1,6 +1,0 @@
-package main.java.com.example.DecimalZipProblem;
-
-public interface Solution
-{
-  public int solution(int A, int B); 
-}
