+++ src/main/java/com/example/SplitTextIntoEqualLengthSubstrings/Main.java.bak
