+++ src/main/java/com/example/city_connection/Solution.java
@@ -1,0 +1,6 @@
+package com.codility.city_connection;
+
+public interface Solution 
+{
+	public int[] solution(int[] T);
+}
